requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain
langchain-chroma
langchain-community
